affFeat

+++ api/src/main/java/pl/kolendateam/dadcard/feats/dto/FeatsDTO.java
@@ -10,18 +10,18 @@ import pl.kolendateam.dadcard.feats.entity.FeatsTypeEnum;
 @AllArgsConstructor
 public class FeatsDTO {
 
-    public short id;
-    public String featName;
-    public FeatsTypeEnum featsType;
-    public PrerequisiteDTO prerequisite;
-    public String description;
+  public int id;
+  public String featName;
+  public FeatsTypeEnum featsType;
+  public PrerequisiteDTO prerequisite;
+  public String description;
 
-    public FeatsDTO(Feats feats){
-        this.id = feats.getId();
-        this.featName = feats.getFeatName();
-        this.featsType = feats.getFeatsType();
-        this.prerequisite = MapperPrerequisiteDTO.toPrerequisiteDTO(feats.getPrerequisite());
-        this.description = feats.getDescription();
-    }
-    
+  public FeatsDTO(Feats feats) {
+    this.id = feats.getId();
+    this.featName = feats.getFeatName();
+    this.featsType = feats.getFeatsType();
+    this.prerequisite =
+      MapperPrerequisiteDTO.toPrerequisiteDTO(feats.getPrerequisite());
+    this.description = feats.getDescription();
+  }
 }
